backend complete

+++ backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler');
 
-const Goal = require ('../model/goalModel');
+const Goal = require ('../models/goalModel');
+const User = require ('../models/userModel')
 
 const getGoals = asyncHandler(async(req,res) =>{
 
-    const goals = await Goal.find()
+    const goals = await Goal.find( {user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -14,7 +15,8 @@ const setGoal = asyncHandler(async(req,res) =>{
         throw new Error('Please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -27,6 +29,17 @@ const updateGoal = asyncHandler(async(req,res) =>{
         res.status(400)
 
         throw new Error('Goal not found');
+    }
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //make sure the logged in user matches the goal user 
+    if(goal.user.toString() != user.id){
+        res.status(401)
+        throw new Error("User not authorized")
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,
         {
@@ -44,6 +57,18 @@ const deleteGoal = asyncHandler(async(req,res) =>{
         res.status(400)
 
         throw new Error ('Please enter existing goal')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //make sure the logged in user matches the goal user 
+    if(goal.user.toString() != user.id){
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
